Создает и импорирует модули
