+++ index.html
@@ -230,6 +230,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
   <!-- Подключение скриптов -->
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
+  <script src="js/utils/utils.js" type="module"></script>
 </body>
 </html>
