+++ js/utils/utils.js
@@ -1,0 +1,8 @@
+// Получение случайного числа
+
+const getRandomPositiveInt = (min, max) =>
+  min >= 0 && min < max
+    ? ~~(min + Math.random() * (max - min + 1))
+    : null;
+
+export {getRandomPositiveInt};
